Make VersionValidator produce better suggestions
The Validator was renamed to ContinuesVersionsValidator
to better describe the purpose of this validator.

Secondly suggested versions are no longer produced using and output argument.
But via a separate getter method.

Lastly the algorithm was completely rewritten and follows the Clean Code principles.

Note: The suggested versions did not take existing versions into account.
Now, instead if a newer major or minor version already exists it only allows a patch release
for the bounded minor version. Otherwise all possible suggestions are accepted.